Defer to Resque.constantize when Resque.constantize is defined
After further investigation, it was found that Resque 1.25.2 does not
define Resque.constantize.  Furthermore the definition of constantize in
the PushPop plugin and extension of the ClassMethods methods was
overriding Resque.constantize.  We want the best of both worlds, so if
Resque.constantize is defined, we will call it.  If not, we will use the
_constantize method that is borrowed from ActiveSupport.